UI changes- modified profile component, ai-resume generate component
