+++ src/app/pages/forms/ai-resume/ai-resume.component.html
@@ -1,0 +1,33 @@
+<div class="row">
+  <div class="col-lg-12">
+    <nb-card>
+      <nb-card-header>Create AI Resume</nb-card-header>
+      <nb-card-body>
+        <div class="container">
+          <div class="row">
+              <div class="col-4">
+                <form><textarea
+                class="form-control bg-light"
+                id="professionalSummary"
+                rows="15"
+                
+                placeholder="Write Job Description"
+              ></textarea></form>
+              </div>
+              <div class="col-2">
+                <button  nbButton [status]=status outline [size]="size">
+                  AI Generate 
+                </button>
+              </div>
+              <div class="col-4">
+                <div>--- placeholder for viewer ---</div>
+              </div>
+          </div>
+      </div>
+        
+      </nb-card-body>
+    </nb-card>
+
+    
+  </div>
+</div>
