+++ src/app/pages/forms/forms-routing.module.ts
@@ -6,6 +6,7 @@ import { FormInputsComponent } from './form-inputs/form-inputs.component';
 import { FormLayoutsComponent } from './form-layouts/form-layouts.component';
 import { DatepickerComponent } from './datepicker/datepicker.component';
 import { ButtonsComponent } from './buttons/buttons.component';
+import { AiResumeComponent } from './ai-resume/ai-resume.component';
 
 const routes: Routes = [
   {
@@ -27,6 +28,10 @@ const routes: Routes = [
       {
         path: 'buttons',
         component: ButtonsComponent,
+      },
+      {
+        path: 'ai-resume',
+        component: AiResumeComponent,
       },
       {
         path: 'datepicker',
