+++ src/app/pages/layout/stepper/profile/profile.component.scss
@@ -1,0 +1,8 @@
+:host ::ng-deep nb-stepper .step-content {
+  text-align: center;
+
+  button {
+    cursor: pointer;
+    margin: 0.5rem;
+  }
+}
