+++ src/app/pages/layout/stepper/stepper.component.ts
@@ -1,8 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormGroup } from '@angular/forms';
-import { Router } from '@angular/router';
-
-import { UserAPI } from '../../../service/api/user-api.service';
+import { UntypedFormBuilder, UntypedFormGroup, Validators } from '@angular/forms';
 
 @Component({
   selector: 'ngx-stepper',
@@ -11,109 +8,36 @@ import { UserAPI } from '../../../service/api/user-api.service';
 })
 export class StepperComponent implements OnInit {
 
-  personalDetails: FormGroup;
-  experience: FormGroup;
-  project: FormGroup;
-  education: FormGroup;
-  certifications: FormGroup;
-  skills: FormGroup;
-  professionalSummary: FormGroup;
-  userId: string
+  firstForm: UntypedFormGroup;
+  secondForm: UntypedFormGroup;
+  thirdForm: UntypedFormGroup;
 
-  constructor(private fb: FormBuilder, private userAPI: UserAPI, private router: Router) {
+  constructor(private fb: UntypedFormBuilder) {
   }
 
   ngOnInit() {
-    this.personalDetails = this.fb.group({
-      username: [''],
-      email: [''],
-      phone: [''],
-      address1: [''],
-      address2: [''],
-      city: [''],
-      state: [''],
-      zip: [''],
-      linkedinLink: [''],
-      portfolioLink: [''],
+    this.firstForm = this.fb.group({
+      firstCtrl: ['', Validators.required],
     });
 
-    this.experience = this.fb.group({
-      position: [''],
-      employer: [''],
-      location: [''],
-      startDate: [''],
-      endDate: [''],
-      currentJob: [''],
-      companyLink: [''],
-      description: [''],
+    this.secondForm = this.fb.group({
+      secondCtrl: ['', Validators.required],
     });
 
-    this.project = this.fb.group({
-      name: [''],
-      link: [''],
-      employer: [''],
-      description: [''],
-      startDate: [''],
-      endDate: [''],
-    });
-
-    this.education = this.fb.group({
-      degreeName: [''],
-      school: [''],
-      location: [''],
-      major: [''],
-      minor: [''],
-      startDate: [''],
-      endDate: [''],
-      grade: [''],
-      description: [''],
-    });
-
-    this.certifications = this.fb.group({
-      name: [''],
-      issuer: [''],
-      startDate: [''],
-      endDate: [''],
-      description: [''],
-    });
-
-    this.skills = this.fb.group({
-      name: [''],
-    });
-
-    this.professionalSummary = this.fb.group({
-      professionalSummary: [''],
+    this.thirdForm = this.fb.group({
+      thirdCtrl: ['', Validators.required],
     });
   }
 
-  async submitForms() {
-    try {
-      // Combine professional summary with personal details
-      const personalDetailsWithSummary = {
-        ...this.personalDetails.value,
-        professionalSummary: this.professionalSummary.value.professionalSummary,
-        password: 'password',
-        role: 'USER'
-      };
-  
-      // Save personal details including professional summary
-      const userDetailsResponse = await this.userAPI.saveUserDetails(personalDetailsWithSummary).toPromise();
-      this.userId = userDetailsResponse.data.id;
-  
-      // Save other details using the obtained user ID
-      await this.userAPI.saveExperience(this.experience.value, this.userId).toPromise();
-      await this.userAPI.saveEducation(this.education.value, this.userId).toPromise();
-      await this.userAPI.saveProjects(this.project.value, this.userId).toPromise();
-      await this.userAPI.saveSkills(this.skills.value, this.userId).toPromise();
-      await this.userAPI.saveCertifications(this.certifications.value, this.userId).toPromise();
+  onFirstSubmit() {
+    this.firstForm.markAsDirty();
+  }
 
+  onSecondSubmit() {
+    this.secondForm.markAsDirty();
+  }
 
-      console.log("userID: ", this.userId)
-  
-      this.router.navigateByUrl(`/pages/layout/stepper/profile?userId=${this.userId}`);
-    } catch (error) {
-      console.error('Error occurred while saving user details:', error);
-      // Handle error
-    }
+  onThirdSubmit() {
+    this.thirdForm.markAsDirty();
   }
 }
