+++ src/app/pages/pages-menu.ts
@@ -1,44 +1,43 @@
-import { NbMenuItem } from '@nebular/theme';
+import { NbMenuItem } from "@nebular/theme";
 
 export const MENU_ITEMS: NbMenuItem[] = [
   {
-    title: 'Home',
-    icon: 'home-outline',
-    link: '/pages/dashboard',
+    title: "Home",
+    icon: "home-outline",
+    link: "/pages/dashboard",
     home: true,
-    hidden: true
+    hidden: true,
   },
   {
-    title: 'FEATURES',
+    title: "FEATURES",
     group: true,
   },
   {
-    title: 'Create',
-    icon: 'layout-outline',
+    title: "Create",
+    icon: "layout-outline",
     children: [
       {
-        title: 'Resume',
-        link: '/pages/layout/stepper',
+        title: "Resume",
+        link: "/pages/layout/stepper/profile",
       },
       {
-        title: 'Cover Letter',
-        link: '',
-      }
+        title: "Cover Letter",
+        link: "",
+      },
     ],
   },
   {
-    title: 'AI',
-    icon: 'keypad-outline',
+    title: "AI",
+    icon: "keypad-outline",
     children: [
       {
-        title: 'AI Resume',
-        link: '',
+        title: "AI Resume",
+        link: "/pages/forms/ai-resume",
       },
       {
-        title: 'AI Cover Letter',
-        link: '',
-      }
+        title: "AI Cover Letter",
+        link: "",
+      },
     ],
   },
- 
 ];
